Updated HITW Scores
Added 4th and 5th places to the score using the above standard; but if something else needs to change or if the array is assumed to be 3 long this may cause errors (sorry!).
Additionally inflated coin values so that coin payout is greatly increased. Still need to add elimination coins and extra coins as time passes when Alex finishes that bit of code.